rearranged to ducks and updated all imports
