--- js/CloudProviders.js
@@ -1,3 +1,0 @@
-const CloudProviders = [{ value: 'Amazon (AWS)' }, { value: 'Google (GKE)' }];
-
-export default CloudProviders;
