+++ js/data/CloudProviders.js
@@ -1,0 +1,6 @@
+const CloudProviders = [
+  { value: 'Amazon (AWS)' },
+  { value: 'Google (GKE)' }
+];
+
+export default CloudProviders;
