+++ js/reducers/rootReducer.js
@@ -1,7 +1,7 @@
 import { combineReducers } from '@reduxjs/toolkit';
 
-import clustersReducer from './ClustersSlice';
-import podsReducer from './PodsSlice';
+import clustersReducer from '../components/Clusters/ClustersSlice';
+import podsReducer from '../components/Pods/PodsSlice';
 
 const rootReducer = combineReducers({
   Clusters: clustersReducer,
